fix with int return

+++ p5_function_v1.sql
@@ -5,7 +5,7 @@ RETURNS FLOAT
 DETERMINISTIC 
 NO SQL
 BEGIN
-    DECLARE result INT;
+    DECLARE result FLOAT;
     SET result = first_num / second_num;
     RETURN result;
 END //
